requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
appdirs==1.4.4
ase==3.22.1
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.1.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cycler==0.11.0
Cython==3.0.0
#dgl==0.7.2
docker-pycreds==0.4.0
docopt==0.6.2
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.17.3
google-auth-oauthlib==0.4.6
googledrivedownloader==0.4
grpcio==1.57.0
h5py==3.8.0
idna==3.4
importlib-metadata==4.13.0
isodate==0.6.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.4
littleutils==0.2.2
llvmlite==0.39.1
lmdb==1.4.1
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.5.3
multidict==6.0.4
networkx==2.6.3
numba==0.56.4
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
ogb==1.3.1
outdated==0.2.2
packaging==23.1
pandas==1.3.5
pathtools==0.1.2
Pillow==9.5.0
pipreqs==0.4.13
protobuf==3.20.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyDeprecate==0.3.0
# pyg-lib==0.2.0+pt113cu117
pyparsing==3.1.1
python-dateutil==2.8.2
python-louvain==0.16
pytorch-lightning==1.3.1
pytz==2023.3
PyYAML==5.4.1
rdflib==6.3.2
rdkit-pypi==2022.9.5
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
sentry-sdk==1.29.2
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.4.1
threadpoolctl==3.1.0
#torch==1.13.1+cu117
#torch-cluster==1.6.1+pt113cu117
#torch-geometric==2.3.1
#torch-scatter==2.1.1+pt113cu117
#torch-sparse==0.6.17+pt113cu117
#torch-spline-conv==1.2.2+pt113cu117
#torchaudio==0.13.1+cu117
torchmetrics==0.5.0
#torchvision==0.14.1+cu117
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.4
wandb==0.15.8
Werkzeug==2.2.3
yarg==0.1.9
yarl==1.9.2
zipp==3.15.0
